refactor(auth): パスキーlibの配置を整理する
