+++ src/storybook/preview.tsx
@@ -6,7 +6,7 @@ import { sb } from 'storybook/test'
 import '../index.css'
 
 sb.mock('../features/auth/lib/auth-client.ts')
-sb.mock('../features/auth/lib/passkey.webauthn-support.ts')
+sb.mock('../features/auth/lib/passkey/webauthn-support.ts')
 
 export default definePreview({
   parameters: {
